backend/requirements.txt
djangorestframework==3.14.0
PyJWT==2.8.0